Added Comments

+++ console/migrations/m130524_201442_init.php
@@ -20,7 +20,7 @@ class m130524_201442_init extends Migration {
             'email' => $this->string()->notNull()->unique(),
             'is_active' => $this->boolean()->notNull()->defaultValue(1),
             'last_login' => $this->integer()->null(),
-            'role_id' => $this->integer()->unsigned()->defaultValue(1),
+            'role_id' => $this->integer(10)->unsigned()->defaultValue(1),
             'created_at' => $this->integer()->notNull(),
             'updated_at' => $this->integer()->notNull(),
                 ], $tableOptions);
